reorganized wallet parts since it's not module anymore
