+++ front/src/app/app.module.ts
@@ -28,18 +28,18 @@ import {ContractSourceComponent} from './components/contract-source/contract-sou
 import {TokenTxsComponent} from './components/token-txs/token-txs.component';
 import {TokenHoldersComponent} from './components/token-holders/token-holders.component';
 
-import {WalletComponent} from './modules/wallet/wallet/wallet.component';
-import {WalletMainComponent} from './modules/wallet/wallet-main/wallet-main.component';
-import {WalletCreateComponent} from './modules/wallet/wallet-create/wallet-create.component';
-import {WalletAccountComponent} from './modules/wallet/wallet-account/wallet-account.component';
-import {DeployerComponent} from './modules/wallet/deployer/deployer.component';
-import {SenderComponent} from './modules/wallet/sender/sender.component';
-import {InteractorComponent} from './modules/wallet/interactor/interactor.component';
+import {WalletComponent} from './scenes/wallet/wallet.component';
+import {WalletMainComponent} from './scenes/wallet-main/wallet-main.component';
+import {WalletCreateComponent} from './scenes/wallet-create/wallet-create.component';
+import {WalletAccountComponent} from './scenes/wallet-account/wallet-account.component';
+import {DeployerComponent} from './components/deployer/deployer.component';
+import {SenderComponent} from './components/sender/sender.component';
+import {InteractorComponent} from './components/interactor/interactor.component';
 /*SERVICES*/
 import {ApiService} from './services/api.service';
 import {CommonService} from './services/common.service';
 import {LayoutService} from './services/layout.service';
-import {WalletService} from './modules/wallet/wallet.service';
+import {WalletService} from './services/wallet.service';
 import {MetaService} from './services/meta.service';
 /*GUARDS*/
 import {WalletGuard} from './guards/wallet.guard';
