+++ front/src/app/scenes/wallet-account/wallet-account.component.ts
@@ -1,13 +1,13 @@
 /*CORE*/
 import {Component, OnDestroy, OnInit} from '@angular/core';
 /*SERVICES*/
-import {WalletService} from '../wallet.service';
-import {MetaService} from '../../../services/meta.service';
-import {CommonService} from '../../../services/common.service';
+import {WalletService} from '../../services/wallet.service';
+import {MetaService} from '../../services/meta.service';
+import {CommonService} from '../../services/common.service';
 /*MODELS*/
-import {Address} from '../../../models/address.model';
+import {Address} from '../../models/address.model';
 /*UTILS*/
-import {META_TITLES} from '../../../utils/constants';
+import {META_TITLES} from '../../utils/constants';
 
 @Component({
   selector: 'app-wallet-account',
