+++ front/src/app/scenes/wallet/wallet.component.scss
