+++ front/src/app/scenes/wallet/wallet.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { WalletComponent } from './wallet.component';
-import {AppModule} from '../../../app.module';
+import {AppModule} from '../../app.module';
 
 describe('WalletComponent', () => {
   let component: WalletComponent;
