+++ front/src/app/scenes/wallet/wallet.component.ts
