+++ front/src/app/services/wallet.service.spec.ts
@@ -1,6 +1,6 @@
 import {TestBed} from '@angular/core/testing';
 import {WalletService} from './wallet.service';
-import {AppModule} from '../../app.module';
+import {AppModule} from '../app.module';
 
 describe('WalletService', () => {
   beforeEach(() => TestBed.configureTestingModule({
